+++ package.json
@@ -13,7 +13,9 @@
     "graphql-compose": "^7.12.3",
     "graphql-compose-mongoose": "^7.3.4",
     "mongoose": "^5.9.2",
-    "mongoose-moment": "^0.1.3"
+    "mongoose-moment": "^0.1.3",
+    "express": "^4.17.1",
+    "express-graphql": "^0.9.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.2",
